requirements.txt

django
djangorestframework
channels
channels-redis
gunicorn
pillow
daphne
whitenoise
psycopg2-binary
dj-database-url
uvicorn
django-environ


